requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0
openai>=1.50.0
streamlit>=1.38.0
chromadb>=0.5.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
pandas>=2.0.0
openpyxl>=3.1.0
